+++ test/utils/index.test.ts
@@ -7,6 +7,7 @@ import {
   safeLoad,
   enableStoreTestingMode,
 } from '../../src';
+import { delayValue } from '../helpers';
 
 enableStoreTestingMode();
 
@@ -111,13 +112,8 @@ describe('rebounce', () => {
 
   const toUpperCase = (input: string) => input.toUpperCase();
 
-  const asyncToUpperCase = (input: string) => {
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        resolve(input.toUpperCase());
-      }, interval());
-    });
-  };
+  const asyncToUpperCase = (input: string) =>
+    delayValue(input.toUpperCase(), interval());
 
   it('works with no timer or rejects', () => {
     const rebouncedToUpperCase = rebounce(asyncToUpperCase);