+++ CHANGELOG.md
@@ -6,6 +6,7 @@ notes:
 - all mapping load functions are rebounced by default
 - if an async store loses all subscriptions and then gains one the mapping load function will always evaluate even if the inputs have not changed
 - can't use stores to hold errors
+- rebounce clear is now called abort
 
 ## 1.0.17 (2023-6-20)